add a test and rename some files
